fix: Fix rerank_documents node in sub_section_writer & qna_agent
